+++ preferences.cpp
@@ -1,9 +1,9 @@
 #include "preferences.h"
 
 #ifdef NIX
-void loadCameraSettings(char* filename, double& focalLength, int& exposureMode) {
+void loadCameraSettings(char* filename, float& focalLength, int& exposureMode) {
 #else
-void loadCameraSettings(char* filename, double& focalLength) {
+void loadCameraSettings(char* filename, float& focalLength) {
 #endif
     rapidxml::xml_document<> doc;
     rapidxml::file<> file(filename);
